[Patchset 12] trace_processor: insert ftrace events into trace storage
This CL introduces the trace inserter class which correltates sched
switch events to compute durations using data in TraceStorage.
TraceParser utilises the inserter as an indirect way of inserting into
the trace storage.

Bug: 80416541

+++ src/trace_processor/trace_parser.h
@@ -40,13 +40,13 @@ class TraceParser {
   void ParseNextChunk();
 
  private:
-  void ParsePacket(const uint8_t* data, uint64_t length);
-  void ParseFtraceEventBundle(const uint8_t* data, uint64_t length);
-  void ParseFtraceEvent(uint32_t cpu, const uint8_t* data, uint64_t length);
+  void ParsePacket(const uint8_t* data, uint32_t length);
+  void ParseFtraceEventBundle(const uint8_t* data, uint32_t length);
+  void ParseFtraceEvent(uint32_t cpu, const uint8_t* data, uint32_t length);
   void ParseSchedSwitch(uint32_t cpu,
                         uint64_t timestamp,
                         const uint8_t* data,
-                        uint64_t length);
+                        uint32_t length);
 
   BlobReader* const reader_;
